workflows: publish to crates.io on release
Adds a workflow triggered by the publishing of a GitHub release which
publishes a crate to crates.io. While the repo is named `libprio-rs`,
Rust naming conventions frown on both the prefix and the suffix: `lib`
is implied by it being a library crate and `-rs` is implied by it being
a Rust crate at all
(https://rust-lang.github.io/api-guidelines/naming.html,
https://doc.rust-lang.org/stable/rust-by-example/crates/lib.html).
Accordingly, the crate is simply named `prio`.

This also adds instructions to the README for cutting releases in GitHub
and also fixes the existing status icons.