finite_field: Define trait FieldElement and add more fields
Defines a new trait, FieldElement, that specifies the API currently
implemented by Field, as well as ome additional features.

Adds a new macro, make_field!(), which wraps a FieldParameters into an
implementation of FieldElement.

Changes the implementation of the 32-bit field Field so that it is
constructed using make_field!().

Adds three additional fields: Field64, Field80, and Field126.