Implement an iterative FFT algorithm
This change adds an alternative algorithm for computing the discrete
Fourier transform. It also adds a new module, benchmarked, for
components of the crate that we want to benchmark, but don't want to
expose in the public API. Finally, it adds a benchmark for comparing the
speed of iterative FFT and recursive FFT on various input lengths.