facilitator: return `Result` from `Server`
Various methods on `Server` used to return `Option<T>`, making it
possible for them to fail but not to indicate why. This means
`prio-server` can't tell the difference between a packet being rejected
due to a decryption error or because the decrypted payload was invalid,
causing frustrating situations as described in [1]. We now return
`Result<T, ServerError>` from `Server`'s methods.

In support of this, this commit also includes some other cleanups:

 - the functions `unpack_proof` and `unpack_proof_mut` now also return
   errors
 - a number of `use module::*` style statements are rewritten to be more
   explicit.

[1] abetterinternet/prio-server#550