Add warnings to field API
The parameters of the 64, 80, and 126 bit fields are subject to change.
This notes this fact in the API. (The 32 bit field is not likely to
change since it's already deployed in Prio v2.)

Also, this corrects the docucomment for `field`, which currently states
that only a 32 bit field is implemented.