pcp: Distinguish between joint and query randomness
The "query randomness" is the random point at which the verifier
evaluates the proof polynomial; the "joint randomness" is whatever
randomness is used by both the prover and verifier. This allows us to
implement proof systems like [BCG+19, Theorem 5.3] in which the verifier
sends the prover a random challenge to use for proof generation.

This commit also cleans up the test code a bit in pcp/types.rs and the
benchmarks in benches/speed_tests.rs.