[WIP] Changes needed for prototype
Add the MeanVarUnsignedVector type, which represents a vector of
L-bit integers. The collector wants to compute the mean and variance of
each element of the vector.

Refactor prng API so that we don't have unwind the entire output for
every call.

Replace the clunky Value::new_with() call with a proper TryFrom that
takes a vector of field elements and "parameters" specified by
Value::Param. (There is also a new call Value::param() for getting the
parameters of an input.)

Add Value::set_leader() call for signaling if an input is the leader's
or a helper's share. (This seems to be useful for certain proof
systems.)

Add an interface for handling Verifier as a vector of field elements.

Move poly_range_check to the polynomial module.