pcp: Add TryFrom for Value from field sequence
The new_with() method on Value allows a new instance of Value to be
constructed from an existing instance and a sequence of field elements.
This is used to construct a Value from a secret share of the underlying
data. However, it requires the caller to already have an instance of
Value on hand, which is usually not the case.

This change drops new_with() in favor of a proper TryFrom supertrait on
Value. The caller provides a vector of field elements as well as a
"parameter", the type of which is an associated type of the particular
implementation of Value.