Use getrandom instead of rand for random numbers
We currently use rand=0.7.3. Support for WASM was dropped in rand=0.8,
so users won't be able to compile libprio-rs if we upgrade this
dependency.

This change aims to nip this problem in the bud by dropping the rand
dependency and replacing it with getrandom, which will support WASM
in the future.