prng: Fix block alignment bug
The Prng structure has a buffer that stores a segment of the keystream.
When it's time to emit the next field element, it reads chunks of this
buffer until a chunk is found that is smaller than the field modulus.
Once the end of the buffer is reached, the buffer is refreshed with the
next segment of the keystream.

The size of the buffer is a multiple of the block size for the
underlying blockcipher (16 bytes). If the size of each field element
chunk does not divide the length of the buffer evenly, then the current
implementation will read passed the end of the buffer.

This changes the length of the buffer so that it's guaranteed to be a
multiple of the length of each field element.