prng: Clear useless bits during rejection sampling
To map a random byte string to a field element, we divide the byte
string into `n`-bit chunks, where `n` is the number of bits needed to
encode the field element. We interpret each chunk as an `n`-bit,
unsigned integer. For each integer, we check if the integer is smaller
than the field modulus: if so, we accept it; otherwise, we reject it and
try the next integer.

This algorithm is fine if `n` is equal to the number of bits needed to
encode the modulus. This isn't always the case, however. In particular,
for Field126, the modulus is a 126-bit prime, but we use 128 bits to
encode field elements. As was pointed out
[here](https://github.com/abetterinternet/prio-documents/issues/14),
this results in rejection happening more frequently than it needs to.

This change clears the extra bits in order to reduce the rate of
ejection. Note that this change is a wire-breaking because impacted
fields will derive a different sequence of field elements for the same
seed. In particular, this impacts Field126, but not Field32, Field64, or
Field80.