prng: Define the StreamCipher trait
Struct Prng uses AES128-CTR as a stream cipher. This commit defines a
trait StreamCipher for stream ciphers and generalizes Prng to use any
implementation of this trait. It also adds an implementation of
StreamCipher for BLAKE3 (https://docs.rs/blake3/1.0.0/blake3). While at
it, this commit reduces the internal API surface for Prng.