pcp: Pass prover randomness to prove()
The prover randomness is now passed to the proof-generation algorithm,
just as the query randomness is passed to the query-generation
algorithm. While at it, this commits adds a callback to Value<F> for
returning the expected length of the prover randomness. It also
normalizes the names of related callbacks for the joint and query
randomness lengths.