Implement prototype of the core PPM protocol
Initial design for the core PPM scheme for Prio. This has not undergone
significant analysis, but is nonetheless useful for prototyping the full
protocol.

This commit also exposes the prng::PrngError in the public API. While
not strictly necessary, it simplifies the task of converting this
internal error. The assumption is that we're going to need this module
in the public API at some point anyway.