Add Serde trait bounds to `field::FieldElement`
We already provide `serde::{Serialize, Deserialize}` implementations for
every implementation of `prio::field::FieldElement`, but we didn't
`express that trait bound on `FieldElement` itself. This makes it tricky
to serialize or deserialize `FieldElement` values in generic functions.

To make this possible, we also change the declaration of various
structs so that the `F: FieldElement` trait bound is on the `impl` and
not the `struct`. See [the Rust API guidelines][rust-api] or [this
clippy issue](clippy) for justification.

[rust-api]: https://rust-lang.github.io/api-guidelines/future-proofing.html#data-structures-do-not-duplicate-derived-trait-bounds-c-struct-bounds
[clippy]: https://github.com/rust-lang/rust-clippy/issues/1689