pcp: Add functionality required for `prio3` VDAF
Adds functionality required for the VDAF. While at it, this simplifies
the interface between `Value` and the PCP methods.

* Define the `ValueParam` trait, which determines the length (in field
elements) of the various inputs and outputs of PCP generation and
verification. This includes things like input length, proof length, and
verifier length that are required for the VDAF.

* Make `ValueParam` a trait bound on `Value::Param` and drop redundant
methods on `Value` (`joint_rand_len`, etc.).

* Add `Value::into_output()` method that transforms a `Value` instance
into an aggregatable output share. This functionality is required for
the VDAF.

* Add `Gadget::calls()` callback so that an instance of a gadget
describes how many times it expects to be called in an arithmetic
circuit. This information is needed to construct all current circuits
anyway.

* Remove `Value::valid_gadget_calls()` as this callback is no longer
needed.