pcp: Align API with draft-patton-cfrg-vdaf-00
Replace `Value` with trait `Type`, implementations of which provide the
FLP functionality described in the spec. In particular, each `Type`
implementation specifies how measurements are encoded and validated. The
FLP system is now implemented as a set of derived methods on a generic
`Type`. However, the type itself does not subsume the input that is
validated.

Shifting the API boundary this way makes it easier to integrate the FLP
into `prio3`.