vdaf::prio3: Use measurement count as aggregation parameter
Changes the type of the aggregation parameter for `Prio3` from `()` to
`usize`. This allows the number of measurements to be used for the
aggregation parameter, which in turn allows this number to be used the
computation of the aggregate result. We can now use prio3 to directly
express things like arithmetic mean, standard deviation, and so on.

The mapping from the aggregation parameter and the finite field
representation of the result to the final result is expressed as a trait
bound. One consequence of this changes it that we will need a unique
result type for each type so that different types can compute this map
differently.