pcp: Properly validate inputs of `Sum::encode()`
If `self.bits == 64` then this method will panic. The behavior that we
want is that encoding should succeed if `2^self.bits` can be encoded as
a field element. This change has performs this check in the constructor.
Also, the constructor now computes the maximum representable integer
than can be represented so that the encoding method can check this
properly.