prio3: Store (compressed) input share instead of output share
The Aggregator must store its output share for the duration of the
validation process. In the PPM protocol, the Helper may end up
serializing this state and sending it encrypted in its HTTP response. In
general, it is more communication efficient to send the seed used to
generate the input share rather than the output share.

This changes the internal structure of `Prio3PrepareStep` so that the
(compressed) input share is stored instead of the output share. While at
it, this changes the interface of `Type::truncate` so that it consumes
its input. This avoids a copy in prio3.