Implement codec traits on `prio3` messages
Adds `Encode` and `Decode` implementations for the `prio3` VDAF protocol
messages, as well as the `Share` and `AggregateShare` types defined in
module `vdaf`. We use the existing `Prio3VerifyParam` and
`Prio3PrepareStep` as the decoding parameters for `Prio3InputShare` and
`Prio3PrepareMessage`, respectively, adding fields where necessary to
provide the context required to decode messages from the wire.

Since we now do all encoding and decoding of `FieldElement`
implementations via the new traits, we can delete
`FieldElement::try_from_reader`.