Remove helper methods on `vdaf::Aggregator`
Remove the methods `Aggregator::{prepare_start, prepare_next,
prepare_finish}`. Using these methods correctly requires the caller to
know statically how many rounds a VDAF requires, because calling
`prepare_next` on a finished preparation or `prepare_finish` on one with
remaining rounds is an error. In order to write a PPM implementation
that is generic over trait `vdaf::Aggregator`, we would have to
introduce a generic means to discover how many rounds a VDAF needs or
whether the `Aggregator::PrepareStep` is on the last round.

The alternative is for a generic implementation to use
`Aggregator::prepare_step`, which returns an enumeration telling the
caller what to do next. Exposing two interfaces for driving the prepare
protocol is more complexity and risk than we need, so we will force
callers to only use one.

This commit also adopts `vdaf::run_vdaf{_prepare}` in `vdaf/poplar1`'s
tests, since the existing tests depended on the helper methods. This
does introduce a signficant performance regression in the unit tests,
because `vdaf::run_vdaf` makes a point of encoding and decoding
messages. With our `ToyIdpf`, this involves encoding and decoding
several hundred thousands of `Field128` values, so `test_poplar1` takes
about 20 seconds on my desktop. Exercising encode/decode of `poplar1`
messages also revealed a bug in `Poplar1InputShare::decode` that this
commit fixes.