Add encoding parameter, allow parameters to be dropped if they are ().
We allow parameters to be dropped by creating a second,
non-parameterized trait, and providing a blanket implementation of the
unparameterized trait that translates the unparameterized methods to the
parameterized methods.

The encoding parameter will practically be used in some PPM messages to
allow the required authentication key to be used as part of the encoding
logic.