bump `aes` and use `ctr`
To get the new version of `cmac` working, I had to bump to the latest
version of `aes`, and also add a dependency on crate `ctr`, as
`Aes128Ctr` is no longer exported by crate `aes`. Using `Ctr64BE` makes
existing tests pass. The documentation clarifies that "64" refers to
[the counter size](https://docs.rs/ctr/latest/ctr/type.Ctr64BE.html),
which must be what `Aes128Ctr` defaulted to.