Switch from Encode/Decode to Into/From<Vec<u8>> for shares.
This is better because we prefer to encode/decode these messages in a
non-streaming fashion (i.e. the size of the buffer is an implicit
parameter to decoding), whereas Encode/Decode must work in a streaming
fashion (where there may be additional values following the "first"
value, so decoding needs to know when to stop).