vdaf: Have prepare_init() output the first-round prep message
Modifies the Aggregator API so that `prepare_init()` outputs the first
round Prepare message. This is less general, since it excludes VDAFs
that may have 0 rounds of communication during the Prepare phase. The
upside is that it eliminates a number of rough edges in both the API and
the implementation. For example, there is never a need to serialize the
initial state prior to producing the first-round message.