Compare FieldElement via Montgomery representation
This changes field element comparison (and hashing) to use Montgomery
representations directly, rather than converting to the natural
representation first. The correctness of this change is dependent on an
invariant that the integer representing a field element is never greater
than the modulus, p, so comments are added throughout to document it.
The field element tests are expanded to exercise the Hash
implementation.