prio3: Move agg result definition to `Type`
Prio3's aggregate result is generic, except that we require that it
implements `TryFrom<AggregateShare<T::Field>>` when unsharding a set of
agg shares. This is a bit awkward, since not every Prio3 instance will
decode the agg result the same way. Thus it is necessary for the
instance to define the type of the agg result and implement the above
`TryFrom<...>`.

This change requires this type to be defined by the underlying `Type`.
This is more natural, since `Type` also defines the measurement type.
This also allows us to do away with the `Prio3Result` and
`Prio3ResultVec` types, which aren't going to be applicable to every
Prio3 instance.