Update Prio3 constructors to improve generated docs.
Previously, the docs for the constructors would show up on the `Prio3`
type, which was confusing since they were all named `new`. Now, the
constructors are free functions with descriptive names. The module
documentation has been updated to reference these functions to provide
readers a hint as to how to construct the various Prio3 VDAF
instantiations.