flp: Fix out-of-bounds error in FFT-based gadget poly computation
When constructing a gadget polynomai, the proof generation algorithm
`Type::prove()` passes to `gadget.call_poly()` the remainder of the
proof string buffer. One of these gadgets, `PolyEval`, uses the length
of this buffer to iterate over an internally allocated buffer. This
leads to an out-of-bounds error when the buffers don't have the same
length.

To address this, `Type::prove()` has been modified in order to pass a
buffer of the correct length to `gadget.call_poly()`. In addition, each
gadget has been modified to strictly enforce the size of the buffer.