Don't unify "real" and dev dependencies
This commit moves the library and `binaries` targets to `resolver = "2"`
so that `dev-dependencies` don't get pulled in when building non-test
configurations of the `prio` library. Up until now, the crate releases
were enabling feature `test-vector`, unnecessarily including the
`test_vector` module and its dependencies. See [1] for details on
dependency resolution with the new resolver.

[1]: https://github.com/rust-lang/cargo/issues/4866