Run CI using both MSRV and stable Rust toolchains
We specify a minimum supported Rust version of 1.58 in `Cargo.toml`, but
were not guaranteeing that crate `prio` actually can build with that
toolchain, instead using whatever the Github Actions Ubuntu runner
happens to have. We now use a matrix to run tests and builds with both
the explicit MSRV, and whatever the current stable toolchain is.

Resolves #284