AggregateShare: switch to TryFrom<&[u8]>, Into<Vec<u8>>. (#335)
Previously, these were written in terms of transformations to and from
references to sequences of bytes. While From & Into can be used in that
way, this doesn't fit with their normal usage, which is to consume some
type and transform it into another type (in this case, consume the
aggregate share and turn it into an owned vector of its serialization,
or vice versa). More practically, specifying things in terms of
references required specifying a "where" clause which effectively became
infectious, needing to be copied to every place the Vdaf trait was
referenced.

Also update OutputShare to specify these traits; they were already
implemented and tested.