Generalize `CountVec` into `SumVec`
The new FLP type, `SumVec`, is used to compute the element-wise sum of a
vector of integers in range `[0, 2^bits)`. Each element of the vector is
encoded as a bit-vector then concatenated together into the FLP input.
The prover then proves that each element of the flattened vector is a
bit in the desired range.

Using `SumVec` is equivalent to using as many `Sum` proofs as there are
elements of the vector. Combining them into one proof improves
efficiency by reducing overall communication cost.

The valididty circuit is identical to `CountVec`; only the chunk size
has changed. In fact, `SumVec` with `bits = 1` should be wire-compatible
with `CountVec`.