vdaf: Simplify trait bound on `Vdaf::PublicShare`
Using a reference to the `Vdaf` as the decoding parameter, as the
`ParameterizedDecode::decode_with_param()` method already expects a
reference. Note that this is only necessary in situation where we want
the decoding parameter to be a tuple and we want to avoid cloning the
tuple item. I.e., the bound on `Vdaf::InputShare` requires a lifetime.