poplar1: Add speed tests for sharding and preparation
The computationally intensive parts of Poplar1 are sharding and
initializing preparation (requires evaluating the IDPF shares). Add
speed tests for both of these computations.

The speed of the preparation step is dominated by the length and
distribution of the candidate prefixes. As such, it is necessary to pick
a distribution that is realistic. Following [BBCG+21], we sample the
measurements from a Zipf distribution and derive the candidate prefixes
from the prefix tree for these measurements.

While at it, add an additional unit test for `poplar1` for exercising
the end-to-end heavy hitters functionality.