Add nonce to `Client::shard()`
In VDAF-04, Prio3 sharding depends on the nonce, so add a nonce to
`Client::shard()`.

The length of the nonce is determined by the VDAF. Accordingly, add a
const generic `NONCE_SIZE` to `Client` and `Aggregator`, both of which
require a nonce. While at it, rename the `L` const generic for
`Aggregator` to `VERIFY_KEY_SIZE` in order to better align with the
spec.