Align domain separation with draft (#456)
Update the `Prg` API by splitting the `info` string into `custom` and
`binder`, as required by VDAF-04. Update `Prio3` and `Poplar1` to use
the new API.

While at it, format each `custom` and `binder` string as specified in
the draft. This is a breaking change, so ignore the test vector tests as
well.