Poplar1: check underflow in Poplar1AggregationParam::try_from_prefixes.
Previously, if an empty IdpfInput was provided, integer underflow would
be encountered on the modified lines; this would cause debug builds to
panic, and almost surely return the wrong error (except on platforms
where usize is 16 bits, in which case very strange behavior was likely).

Also, update the comment about possible errors to include a few
undocumented errors.