`Prio2`: Remove proof validation state
The current implementation of `Prio2` constructs an object called
`ValidationMemory` that includes the memory for the recursive FFT
algorithm and pre-computed roots of unity used for FFT and rejecting
query randomnesses `eval_at`. (It's not secure to query the proof at a
root of unity.) This memory can be removed by making the following
changes:

* Use the iterative FFT algorithm
* To do the root-of-unity check, check if `eval_at`, when raised to the
  power of the proof polynomial length, is equal to `1`. If not, then
  `eval_at` is not a root of unity.

According to the microbenchmarks, removing this memory and making these
changes results in a 20% speed up for dimension = 1,000, 16% for 10,000,
and 10% for 100,000. (Smaller inputs fair even better.)