fft: Add SIMD variant of `discrete_fourier_transform()`
There are a number of situations where FFT is run on multiple inputs of
the same length. This function includes some arithmetic that ends up
being identical for each input.

This commit adds an alternative method,
`mutli_discrete_fourier_transform()` that deduplicates this arithmetic.
It also integrates into FLP prover and verifier and some of the gadgets.