flp: Flatten the wire polynomials into a single buffer
Modify `Gadget::cal_poly()` by changing the input type from `&[Vec<F>]`
to `&[F]`. The input is now the concatenation of each of the wire
polynomials, rather than allocating a vector for each polynomial. This
changes has two effects, both of which should reduce CPU time of the
prover and verifier:

1. Allocating a single `Vec<F>` rather than a nested `Vec<Vec<F>>`
   removes unnecessary indirection and reduces the number of memory
   allocations. (This is especially noticeable for high-arity gadgets.)

2. The same buffer can be re-used across gadgets.