Remove non-constant-time comparisons of secret values.
This affects the types representing:
  * Input shares
  * Preparation states
  * Output shares
  * Aggregate shares

Mostly, the comparisons were either dropped entirely or updated to be
test-only.  Input shares were instead given a constant-time equality
implementation, as it is believed this is required for DAP
implementations.

PartialEq & Eq implementations are still derived when the "test-util"
feature is enabled. This is to ease testing for users of this library.