Align `Aggregator` trait with spec
`Vdaf::prepare_preprocess` was missing the `&Vdaf::AggregationParam`
argument specified in VDAF. This adds that argument to the trait method
and its implementations. While we're at it, to further align with the
spec, we add new methods `prepare_shares_to_prepare_message` and
`prepare_next`, to align with the spec's `prep_shares_to_prep` and
`prep_next`, as aliases for the existing `prepare_preprocess` and
`prepare_step`. The existing methods are marked as deprecated.

Resolves #670