Prio3: Decouple the algorithm ID from the FLP circuit (#855)
First, replace `Vdaf::ID` with `Vdaf::algorithm_id(&self) -> u32`. 
Instead of defining the algorithm ID as an associated type, define a
method for returning it from an instance of the VDAF. This allows us to
provide more flexibility for variants of (standardized) VDAFs.

Next, move `Type::ID` to `Prio3::new()`. This allows new Prio3 variants
to be defined for a standard circuit.