Improve FLP testing and add tooling to "test-util" (#900)
* Improve FLP testing and add tooling to "test-util"

Implement a new struct, `flp::test_util::FlpTest`, for driving generic
tests for FLPs. This tool subsumes most of the functionality of
`flp_validity_test()`. The irregular length tests have been dropped such
that the `valid()` is no longer required to do bounds checking itself.

Remove `flp_validity_test()` and replace each use with `FlpTest`.

* Don't require "crypto-dependencies" for `random_vector()`

Re-implement `random_vector()` with `XofTurboShake128` so that we don't
have to enable the "crypto-dependencies" feature to use it. The current
implementation uses `SeedStreamAes128`, which is necessarily behind this
feature.