dummy vdaf: make aggregate results predictable (#970)
Users of the dummy VDAF need to be able to predict what it will output
so they can verify that they have executed it correctly. To that end,
summations performed by the dummy VDAF are now done modulo `u8::MAX`.
Additionally, it's nice for aggregations run with distinct aggregation
parameters to yield distinct aggregate results, so we now sum in the
aggregation parameter when unsharding aggregate shares. Finally, we add
`vdaf::dummy::expected_aggregate_result` as a convenience function.