Create module for Szk proof system (#1013)
Szk wraps an FLP proof system and uses an Xof to derive joint random
coins using a method inspired by Prio3. However, unlike Prio3, Szk
is designed to prove validity of an input that is already secret-shared
between two parties. This is necessary for Mastic, where the input
will be shared via a VIDPF instead of linearly.

Co-authored-by: Hannah Davis <hannah.e.davis@seagate.com>