Use generic implementations for prime field operations.
This is a refactor of the FieldParameters struct
to use generic datatypes providing implementations for
primes that fit in one word.

Tests script and documentation parameters were updated to
support the new structure.

Performance for FP32 is twice as fast than before, and no
changes for the FP64 and FP128 fields.