vidpf: Move onehot proof computation out of `eval_next()`
We're currently computing the onehot proof via the following tree
traversal: for each prefix, visit each node along the path from the root
to the node labeled by the prefix. The Mastic spec calls for a
breadth-first traversal of the sub-tree composed of the nodes of the
prefixes.

This commit makes three improvements towards aligning Mastic with the
spec.

First, de-couple the onehot proof computation from `eval_next()`
completely. At the moment, the "latest onehot proof" is stored in the
evaluation state, a value that is dependent on the tree traversal. This
forces us to handle Mastic-specific details in VIDPF. Instead we store
the node proof in the evaluation state, which is independent of the
traversal.

Second, modify `eval()` and `eval_cache()` to return a onehot proof for
the single prefix they traverse.

Third, consolidate the `eval_root()` APIs into a single function.