Various improvements to VIDPF (#1177)
* vidpf: Align terminology with spec

* vidpf: Merge "cs" into correction word

The "cs" field is merely a correction word for the node proof.

* vidpf: Clean up API documentation

Clean up comments a bit and move a couple of things out of the public
API.

* vidpf: Remove `NONCE_SIZE` parameter

This simplifies the API a bit and improves parity with the idpf module.

* vidpf: Improve control flow

Create an iterator over a `VidpfInput` that consolidates some error
handling common to VIDPF generation and evaluation. This allows us to
reduce the amount of error handling necessary and improve control flow
in a few places.

* vidpf: Move onehot proof computation out of `eval_next()`

We're currently computing the onehot proof via the following tree
traversal: for each prefix, visit each node along the path from the root
to the node labeled by the prefix. The Mastic spec calls for a
breadth-first traversal of the sub-tree composed of the nodes of the
prefixes.

This commit makes three improvements towards aligning Mastic with the
spec.

First, de-couple the onehot proof computation from `eval_next()`
completely. At the moment, the "latest onehot proof" is stored in the
evaluation state, a value that is dependent on the tree traversal. This
forces us to handle Mastic-specific details in VIDPF. Instead we store
the node proof in the evaluation state, which is independent of the
traversal.

Second, modify `eval()` and `eval_cache()` to return a onehot proof for
the single prefix they traverse.

Third, consolidate the `eval_root()` APIs into a single function.

* vidpf: Encapsulate output of `eval_next()`

Rename `VidpfEvalCache` to `VidpfEvalResult` and have `eval_enxt()`
output it.

* vidpf: Defer negation of weight until after `eval_next()`

This is needed in order to properly compute the path and payload checks.