Mastic clean up (#1187)
* mastic: Remove `verify_key` from tests that don't use it

Some tests in the `mastic` module generate verification keys that aren't
actually used by the test. Incidentally, the same tests were
generating keys of the wrong length.

* mastic: De-duplicate agg share length computation

Add a function that computes the length of the aggregate share in field
elements as a function of the aggregation parameter.

* vidpf: Improve `VidpfPublicShare::encoded_len()`

Avoid iterating over the weights to compute the length of the encoded
public share; just take the length of the first weight and multiply by
the number of correction words.

This computation assumes the length of each weight is equal to the
weight parameter at ever level of the VIDPF tree. This certainly is
true, but add a test to validate this assumption anyway.

* vidpf: Move `eval_prefix_tree_with_siblings()` to `impl<W: VidpfValue>`

This method is currently implemented for `Vidpf<VidpfWeight<F>>`, but it
applies to the more general `Vidpf<W>`.

* vdaf: Remove `domain_separation_tag()` from `Vdaf` trait

This method is used in Prio3 and Poplar1 for domain separation with the
version of the document that specifies them. This version control is not
applicable to future VDAFs defined by future documents.

Remove the method from the trait and add it to implementations of
`Prio3` and `Poplar1`.

* vidpf: Rename `weight_parameter` to `weight_len`

The associated type `ValueParameter` is likely always going to be a
`usize` that expresses the length. In the future we might consider
hardcoding this change in the API.