More Mastic clean up (#1200)
* xof: Have `seed_stream()` take an array

* mastic: Define constants for nonce and seed sizes

* mastic: Remove `bits` field from `Mastic` struct

This is redundant since `Vidpf` now needs to know the bit length.

* Move `szk` module from `flp` to `vdaf::mastic`

Some of the implementation details of SZK are, for the moment, specific
to Mastic. For example, the domain separation tag for each XOF
invocation is prefixed by `b"mastic"`. It therefore makes sense to move
this module to `mastic` until such time we need to refactor in order to
use SZK in multiple places.

* mastic: Update reference