Add methods for adding/subtracting vectors
Add a method `add_assign_vector` that replaces the following zip
pattern:

```
for (x, y) in a.iter_mut().zip(y) {
    *x += *y;
}
```

The method also panics if `b` is shorter than `a`.

Replace the zip pattern with this method wherever we add up vectors.
This helps us ensure that we're not accidentally adding a vector that's
shorter than the other.

Likewise for subtracting vectors.