codec: expose `decode_fixlen_items`
For https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/pull/686, DAP will
need protocol messages with fixed-length vectors in them ([1]). Module
`codec` already had `encode_fixlen_items` in the public API for this,
but the corresponding decode function was private.

This commit renames it to align with the encode function and makes it
public.

[1]: https://datatracker.ietf.org/doc/html/rfc8446#section-3.4