ping pong topo: use longer variable names
I'm working on bringing the changes in #1259 to `main`. While doing
that, I became frustrated with the unnecessarily terse variable names in
this module, so I expanded them to e.g. `prepare_message` instead of
`prep_msg`. I'm breaking that part of the change out so that the
eventual PR that makes real code changes will be easier to digest.

As this changes some fields of exported enums and structs, it's a
breaking change. Variable naming would never normally suffice to justify
an API break, but the rest of the ping pong topo changes that are coming
are backward-incompatible anyway.

Part of #1266